+++ src/main/java/com/moilioncircle/redis/replicator/cmd/impl/DefaultCommand.java
@@ -14,13 +14,13 @@
  * limitations under the License.
  */
 
-package com.moilioncircle.examples.migration.cmd;
+package com.moilioncircle.redis.replicator.cmd.impl;
 
 import com.moilioncircle.redis.replicator.cmd.Command;
 
 /**
  * @author Leon Chen
- * @since 2.4.3
+ * @since 2.5.0
  */
 public class DefaultCommand implements Command {