+++ src/main/java/com/moilioncircle/redis/replicator/util/CRC64.java
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-package com.moilioncircle.examples.util;
+package com.moilioncircle.redis.replicator.util;
 
 /**
  * @author Leon Chen
- * @since 2.4.3
+ * @since 2.5.0
  */
 public class CRC64 {
     public static final long[] LOOKUP_TABLE = {